+++ src/scip/cons_linear.c
@@ -3183,8 +3183,8 @@ SCIP_DECL_SORTINDCOMP(consdataCompVar)
    }
    else
    {
-      unsigned int vartype1 = SCIPvarIsImpliedIntegral(var1) ? 2 : SCIPvarGetType(var1);
-      unsigned int vartype2 = SCIPvarIsImpliedIntegral(var2) ? 2 : SCIPvarGetType(var2);
+      unsigned int vartype1 = SCIPvarIsImpliedIntegral(var1) ? 2 : (unsigned int) SCIPvarGetType(var1);
+      unsigned int vartype2 = SCIPvarIsImpliedIntegral(var2) ? 2 : (unsigned int) SCIPvarGetType(var2);
 
       if( vartype1 < vartype2 )
          return -1;
@@ -3231,8 +3231,8 @@ SCIP_DECL_SORTINDCOMP(consdataCompVarProp)
    else
    {
       /* Weird check but this was the easiest way to keep the old logic when refactoring implied integers */
-      unsigned int vartype1 = SCIPvarIsImpliedIntegral(var1) ? 2 : SCIPvarGetType(var1);
-      unsigned int vartype2 = SCIPvarIsImpliedIntegral(var2) ? 2 : SCIPvarGetType(var2);
+      unsigned int vartype1 = SCIPvarIsImpliedIntegral(var1) ? 2 : (unsigned int) SCIPvarGetType(var1);
+      unsigned int vartype2 = SCIPvarIsImpliedIntegral(var2) ? 2 : (unsigned int) SCIPvarGetType(var2);
 
       if( vartype1 < vartype2 )
       {
